[FLINk-1680] Remove Tachyon test, rename maven module to "flink-fs-tests"
